Perception isn't listed when adding DC checks.
I noticed in the source what appeared to be a typo. Not sure if the previous spelling was for a particular system that had a typo, or if that is what was causing the error for Pathfinder 2e. Should it be intentional this spelling, could it be added that the attribute class be pulled in order to show perception DC checks.